Safari-611.1.10/LayoutTests/js/dom/document-all-typeof-is-object-fold-expected.txt

typeof document.all is never 'object'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testTypeofIsObject() is true
PASS successfullyParsed is true

TEST COMPLETE
